ft_transcendence/ft_trance_spa/pingpong/requirements.txt
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
constantly==23.10.4
cryptography==42.0.7
daphne==4.1.2
Django==5.0.4
django-rest-framework==0.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
pillow==10.3.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.1.0
python-http-client==3.3.7
sendgrid==6.11.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
starkbank-ecdsa==2.2.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
zope.interface==6.4
requests==2.31.0
django-cors-headers==4.3.1
django-channels-jwt==0.0.3
PyJWT==2.8.0



gunicorn 
uvicorn[standard]
httpx